+++ packages/synapse/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { createEventBus, type EventBusConfig, type EventBusEvent, type EventBusState } from './createEventBus'
 export { createSynapse, type Synapse, type SynapseConfig, type SynapseModule } from './createSynapse/index'
 export { type AwaitableSynapse, createSynapseAwaiter, type SynapseAwaiter } from './createSynapseAwaiter'
+export { dehydrateModule, type DehydrateModuleOptions } from './dehydrateModule'